Javascript Synthesizer Code
A simple synth controlled by clicking on the keys displayed, holding each key keeps the note held down with slight colour changes on the keys to indicate which are being played. 

I will be adding the option to use your own keyboard to control the synth and going on to add midi support.